🚑 Critical hotfix
Segmentation fault due to floats not being casted as ints

+++ include/graphlib.h
@@ -89,5 +89,5 @@ char *get_asset_fullpath(char *filename);
 /* mult_free.c */
 void mult_free(int how_much, ...);
 /* print_message.c */
-void print_message(char *str, sfRenderWindow *win, int size, sfVector3f info);
+void print_message(char *str, sfRenderWindow *win, float size, sfVector3f info);
 #endif
